+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org/)
 since `1.7.0`.
 
 ## [Unreleased]
+### Fixed
+- Fixed characters being dropped when color codes were used.
 
 ## [1.16.0-alpha.9] - 2020-04-24
 ### Fixed
@@ -16,7 +18,6 @@ since `1.7.0`.
   [#13](https://github.com/libtcod/libtcod/issues/13)
   [#50](https://github.com/libtcod/libtcod/pull/50)
 - The `TCOD_VSYNC` environment variable was being ignored.
-- Fixed characters being dropped when color codes were used.
 - Malformed UTF-8 strings will no longer hang printing functions.
 
 ## [1.16.0-alpha.8] - 2020-04-03